[cleanup] Move ViewNodeService services in syson-diagram-services
