+++ CHANGELOG.adoc
@@ -16,6 +16,24 @@ As a result, the following GraphQL mutations have been removed `exposeRequiremen
 - https://github.com/eclipse-syson/syson/issues/2045[#2045] [diagrams] The service `ViewNodeService#revealCompartment` has been changed to reveal a compartment only when it is needed.
 + Only reveal a compartment of a node when none of the revealed compartment can display the `targetElement`.
 + If many compartment candidates exist, choose the first free form compartment.
+- [services] As part of the migration to the new diagram service architecture, the following methods have been moved out of `ViewNodeService` (which as been deleted):
++
+** `List<Element> getExposedElements(Element self, EClass domainType, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)` to `DiagramQueryAQLService.getExposedElements(Element self, EClass domainType, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)`
+** `List<Element> getExposedElements(Element self, Element parent, EClass domainType, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)` to `DiagramQueryAQLService.getExposedElements(Element self, Element parent, EClass domainType, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)`
+** `boolean isView(Element self, String viewDefinition, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)` to `DiagramQueryAQLService.isView(Element self, String viewDefinition, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)`
+** `boolean isView(Element self, String viewDefinition, Node selectedNode, IEditingContext editingContext, DiagramContext diagramContext)` to `DiagramQueryAQLService.isView(Element self, String viewDefinition, Node selectedNode, IEditingContext editingContext, DiagramContext diagramContext)`
+** `boolean isHiddenByDefault(Element self, String compartmentName, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)` to `DiagramQueryAQLService.isHiddenByDefault(Element self, String compartmentName, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)`
+** `List<PartUsage> getExposedActors(Element self, EClass domainType, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)` to `DiagramQueryAQLService.getExposedActors(Element self, EClass domainType, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)`
+** `List<PartUsage> getExposedStakeholders(Element self, EClass domainType, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)` to `DiagramQueryAQLService.getExposedStakeholders(Element self, EClass domainType, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)`
+** `List<ReferenceUsage> getExposedSubjects(Element self, EClass domainType, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)` to `DiagramQueryAQLService.getExposedSubjects(Element self, EClass domainType, List<Object> ancestors, IEditingContext editingContext, DiagramContext diagramContext)`
+** `List<Element> getAllReachableRequirements(EObject self)` to `DiagramQueryAQLService.getAllReachableRequirements(EObject self)`
+** `boolean isAncestorOf(org.eclipse.sirius.components.representations.Element parentNodeElement, org.eclipse.sirius.components.representations.Element childNodeElement, DiagramRenderingCache cache)` to `DiagramQueryAQLService.isAncestorOf(org.eclipse.sirius.components.representations.Element parentNodeElement, org.eclipse.sirius.components.representations.Element childNodeElement, DiagramRenderingCache cache)`
+** `boolean showAnnotatingNode(Element self, DiagramContext diagramContext, IEditingContext editingContext)` to `DiagramQueryAQLService.showAnnotatingNode(Element self, DiagramContext diagramContext, IEditingContext editingContext)`
+** `Node revealCompartment(Node selectedNode, Element targetElement, DiagramContext diagramContext, IEditingContext editingContext, Map<org.eclipse.sirius.components.view.diagram.NodeDescription, org.eclipse.sirius.components.diagrams.description.NodeDescription> convertedNodes)` to `DiagramMutationAQLService.revealCompartment(Node selectedNode, Element targetElement, DiagramContext diagramContext, IEditingContext editingContext, Map<org.eclipse.sirius.components.view.diagram.NodeDescription, org.eclipse.sirius.components.diagrams.description.NodeDescription> convertedNodes)`
+** `org.eclipse.syson.diagram.common.view.services.ViewFilterSwitch` to `org.eclipse.syson.diagram.services.utils.ViewFilterSwitch`
+** `org.eclipse.syson.diagram.common.view.services.RevealCompartmentSwitch` to `org.eclipse.syson.diagram.services.utils.RevealCompartmentSwitch`
++
+Also, a new helper method `DiagramQueryAQLService.isNotAncestorOf(...)` has been added to support edge preconditions previously written as `aql:not graphicalEdgeSource.isAncestorOf(graphicalEdgeTarget, cache)`.
 
 === Dependency update
