+++ backend/views/syson-diagram-common-view/src/main/java/org/eclipse/syson/diagram/common/view/edges/AbstractUsageNestedUsageEdgeDescriptionProvider.java
@@ -84,8 +84,10 @@ public abstract class AbstractUsageNestedUsageEdgeDescriptionProvider extends Ab
                 .conditionalStyles(this.createReferenceConditionalEdgeStyle())
                 .synchronizationPolicy(SynchronizationPolicy.SYNCHRONIZED)
                 .targetExpression(AQLConstants.AQL_SELF + "." + this.eReference.getName())
-                .preconditionExpression(AQLConstants.AQL + "not " + org.eclipse.sirius.components.diagrams.description.EdgeDescription.GRAPHICAL_EDGE_SOURCE + ".isAncestorOf("
-                        + org.eclipse.sirius.components.diagrams.description.EdgeDescription.GRAPHICAL_EDGE_TARGET + ", cache)")
+                .preconditionExpression(ServiceMethod.of2(DiagramQueryAQLService::isNotAncestorOf)
+                        .aql(org.eclipse.sirius.components.diagrams.description.EdgeDescription.GRAPHICAL_EDGE_SOURCE,
+                                org.eclipse.sirius.components.diagrams.description.EdgeDescription.GRAPHICAL_EDGE_TARGET,
+                                "cache"))
                 .build();
     }
