+++ backend/views/syson-diagram-common-view/src/main/java/org/eclipse/syson/diagram/common/view/nodes/AbstractNodeDescriptionProvider.java
@@ -31,7 +31,7 @@ import org.eclipse.sirius.components.view.diagram.NodeToolSection;
 import org.eclipse.sirius.components.view.diagram.provider.DefaultToolsFactory;
 import org.eclipse.sirius.components.view.emf.diagram.ViewDiagramDescriptionConverter;
 import org.eclipse.syson.diagram.services.aql.DiagramMutationAQLService;
-import org.eclipse.syson.diagram.common.view.services.ViewNodeService;
+import org.eclipse.syson.diagram.services.aql.DiagramQueryAQLService;
 import org.eclipse.syson.sysml.Element;
 import org.eclipse.syson.util.AQLConstants;
 import org.eclipse.syson.util.AQLUtils;
@@ -115,7 +115,7 @@ public abstract class AbstractNodeDescriptionProvider implements INodeDescriptio
         return this.diagramBuilderHelper.newNodeTool()
                 .name("Show content as Nested")
                 .iconURLsExpression("/icons/full/obj16/ShowTool.svg")
-                .preconditionExpression(ServiceMethod.of4(ViewNodeService.class, ViewNodeService::isView,
+                .preconditionExpression(ServiceMethod.of4(DiagramQueryAQLService.class, DiagramQueryAQLService::isView,
                         Element.class, String.class, Node.class, IEditingContext.class, DiagramContext.class)
                         .aqlSelf(AQLUtils.aqlString(StandardDiagramsConstants.GV_QN), Node.SELECTED_NODE, IEditingContext.EDITING_CONTEXT, DiagramContext.DIAGRAM_CONTEXT))
                 .body(this.diagramBuilderHelper.newDeleteView()
@@ -131,7 +131,7 @@ public abstract class AbstractNodeDescriptionProvider implements INodeDescriptio
         return this.diagramBuilderHelper.newNodeTool()
                 .name("Show content as Tree")
                 .iconURLsExpression("/icons/full/obj16/ShowTool.svg")
-                .preconditionExpression(ServiceMethod.of4(ViewNodeService.class, ViewNodeService::isView,
+                .preconditionExpression(ServiceMethod.of4(DiagramQueryAQLService.class, DiagramQueryAQLService::isView,
                         Element.class, String.class, Node.class, IEditingContext.class, DiagramContext.class)
                         .aqlSelf(AQLUtils.aqlString(StandardDiagramsConstants.GV_QN), Node.SELECTED_NODE, IEditingContext.EDITING_CONTEXT, DiagramContext.DIAGRAM_CONTEXT))
                 .body(this.diagramBuilderHelper.newDeleteView()
