+++ scripts/check-coverage.jsh
@@ -35,12 +35,12 @@ var moduleCoverageData = List.of(
   new ModuleCoverage("syson-application-configuration", 75.0),
   new ModuleCoverage("syson-common-view", 100.0),
   new ModuleCoverage("syson-diagram-common-view", 89.0),
-  new ModuleCoverage("syson-diagram-services", 84.0),
+  new ModuleCoverage("syson-diagram-services", 85.0),
   new ModuleCoverage("syson-direct-edit-grammar", 67.0),
   new ModuleCoverage("syson-form-services", 100.0),
   new ModuleCoverage("syson-model-services", 94.0),
   new ModuleCoverage("syson-representation-services", 100.0),
-  new ModuleCoverage("syson-services", 65.0),
+  new ModuleCoverage("syson-services", 68.0),
   new ModuleCoverage("syson-siriusweb-customnodes-metamodel", 41.0),
   new ModuleCoverage("syson-siriusweb-customnodes-metamodel-edit", 0.0),
   new ModuleCoverage("syson-standard-diagrams-view", 97.0),
